+++ R/list-xyz-tidiers.R
@@ -37,12 +37,9 @@ tidy_xyz <- function(x, ...) {
     )
   }
 
-  d <- as_tibble(x$z)
-  names(d) <- x$y
-  d$x <- x$x
-  pivot_longer(d,
-    cols = c(dplyr::everything(), -x),
-    names_to = "y",
-    values_to = "z"
-  )
+  as_tibble(data.frame(
+    x = x$x,
+    y = rep(x$y, each = length(x$x)),
+    z = as.numeric(x$z)
+  ))
 }